refactor: Issue publication field array
Moved issue publication field compoenent to its associated field array
