+++ src/components/RulesetFormSections/IssuePublicationFieldArray/IssuePublicationField/index.js
