+++ src/components/RulesetFormSections/index.js
@@ -2,7 +2,6 @@ export { default as RulesetInfoForm } from './RulesetInfoForm';
 export { default as ModelRulesetSelection } from './ModelRulesetSelection';
 export { default as PatternTimePeriodForm } from './PatternTimePeriodForm';
 export { default as IssuePublicationFieldArray } from './IssuePublicationFieldArray';
-export { default as IssuePublicationField } from './IssuePublicationField';
 export { default as OmissionFieldArray } from './OmissionFieldArray';
 export { default as CombinationFieldArray } from './CombinationFieldArray';
 export { default as ChronologyFieldArray } from './ChronologyFieldArray';
